refactor: move event-processing classes to event package
