+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/EventMarker.java
@@ -1,9 +1,6 @@
-package io.javaoperatorsdk.operator.processing;
+package io.javaoperatorsdk.operator.processing.event;
 
 import java.util.HashMap;
-
-import io.javaoperatorsdk.operator.processing.event.Event;
-import io.javaoperatorsdk.operator.processing.event.ResourceID;
 
 /**
  * Manages the state of received events. Basically there can be only three distinct states relevant
