+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/AbstractEventSource.java
@@ -1,4 +1,6 @@
-package io.javaoperatorsdk.operator.processing.event;
+package io.javaoperatorsdk.operator.processing.event.source;
+
+import io.javaoperatorsdk.operator.processing.event.EventHandler;
 
 public abstract class AbstractEventSource implements EventSource {
